requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra=tests,dev --output-file=requirements.txt pyproject.toml
#
black==23.7.0
    # via ocr-web (pyproject.toml)
build==0.10.0
    # via pip-tools
cachetools==5.3.1
    # via tox
chardet==5.2.0
    # via tox
click==8.1.3
    # via
    #   black
    #   flask
    #   pip-tools
colorama==0.4.6
    # via tox
distlib==0.3.7
    # via virtualenv
filelock==3.12.2
    # via
    #   tox
    #   virtualenv
flask==2.2.3
    # via ocr-web (pyproject.toml)

isort==5.12.0
    # via ocr-web (pyproject.toml)
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via
    #   flask
    #   ocr-web (pyproject.toml)
langcodes[data]==3.5.0
    # via ocr-web (pyproject.toml)
marisa-trie==1.2.1
    # via language-data
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
mypy-extensions==1.0.0
    # via black
packaging==23.0
    # via
    #   black
    #   build
    #   pyproject-api
    #   pytesseract
    #   tox
pathspec==0.11.2
    # via black
pdf2image==1.16.2
    # via ocr-web (pyproject.toml)
pillow==11.3.0
    # via
    #   ocr-web (pyproject.toml)
    #   pdf2image
    #   pytesseract
pip-tools==7.3.0
    # via ocr-web (pyproject.toml)
platformdirs==3.10.0
    # via
    #   black
    #   tox
    #   virtualenv
pluggy==1.2.0
    # via tox
psycopg2-binary==2.9.10
pyproject-api==1.5.1
    # via tox
pyproject-hooks==1.0.0
    # via build
pytesseract==0.3.10
    # via ocr-web (pyproject.toml)
ruff==0.0.284
    # via ocr-web (pyproject.toml)
tox==4.4.12
    # via ocr-web (pyproject.toml)
virtualenv==20.24.2
    # via tox
werkzeug==2.2.3
    # via
    #   flask
    #   ocr-web (pyproject.toml)
wheel==0.41.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
google-generativeai
